Convert order and position ID types to i32
orders and positions id/instrument_id/owner_id (and positions' joined
market_id) are INTEGER (int4) in Postgres. Switch Order/CreateOrder,
Position/CreatePosition/PositionWithMarketId and the PositionStore id
parameters to i32, dropping the compensating CAST(... AS BIGINT) casts on
those columns (the Currency/quantity casts stay - those columns really are
int4 vs i64 fields, a separate cleanup).

These two stores share a commit because TradeResult/ResolveResult bundle a
CreateOrder and a CreatePosition that trade.rs builds and the test suite
asserts in lockstep, so the .into() bridging in trade.rs can't be split
without an intermediate state that fails to compile. The temporary `as i32`
casts from the instrument commit are removed here.

+++ bot/src/command/leaderboard.rs
@@ -97,7 +97,7 @@ pub async fn run(
         .position_store
         .get_all_positions_with_market_id(&mut tx)
         .await?;
-    let mut positions_by_user: HashMap<i64, Vec<PositionWithMarketId>> = HashMap::new();
+    let mut positions_by_user: HashMap<i32, Vec<PositionWithMarketId>> = HashMap::new();
 
     // Process positions into a HashMap of lists per user.
     for p in positions {
@@ -123,9 +123,7 @@ pub async fn run(
     let mut portfolio_values = Vec::new();
 
     for user in users.into_values() {
-        let positions = positions_by_user
-            .get(&i64::from(user.id))
-            .unwrap_or(&empty);
+        let positions = positions_by_user.get(&user.id).unwrap_or(&empty);
         portfolio_values.push(PortfolioValue::new(
             user,
             &transfers,
